+++ web-components/src/components/cards/PostCard/PostCard.stories.tsx
@@ -2,10 +2,8 @@ import React from 'react';
 import { MenuItem } from '@mui/material';
 import { StoryObj } from '@storybook/react';
 
-import { User } from 'src/components/user/UserInfo';
-
-import { Body } from '../../typography';
 import PostCard from './PostCard';
+import { commonArgs, signer } from './PostCard.mock';
 
 type Story = StoryObj<typeof PostCard>;
 
@@ -14,38 +12,16 @@ export default {
   component: PostCard,
 };
 
-const user: User = {
-  name: 'Odonata',
-  type: 'ORGANIZATION',
-  image:
-    'https://cdn.sanity.io/images/jm12rn9t/staging/8b062589b6d8c6850a78bb13ead51d2f6f32b073-29x9.svg',
-};
-
-const signer: User = {
-  name: 'Odonata',
-  nameRaw: <Body>Odonata</Body>,
-  type: 'ORGANIZATION',
-  image:
-    'https://cdn.sanity.io/images/jm12rn9t/staging/8b062589b6d8c6850a78bb13ead51d2f6f32b073-29x9.svg',
-};
-
 export const Public: Story = {
   args: {
-    title: 'Fall 2023 Update',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    ...commonArgs,
     imgSrc: '/coorong.png',
-    author: user,
-    authorRole: 'admin',
-    timestamp: '2022-01-31T12:34:56Z',
     isPrivate: false,
     isAdmin: false,
-    handleClickShare: () => {
-      console.log('handle share click');
-    },
-    handleClickFile: () => {
-      console.log('handle file click');
-    },
+  },
+  argTypes: {
+    handleClickShare: { action: 'handle share click' },
+    handleClickFile: { action: 'handle file click' },
   },
   render: args => <PostCard {...args} signer={signer} />,
 };
@@ -58,20 +34,31 @@ const adminMenuItems = [
 
 export const Private: Story = {
   args: {
-    title: 'Fall 2023 Update',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-    imgSrc: '/coorong.png',
-    author: user,
-    authorRole: 'admin',
-    timestamp: '2022-01-31T12:34:56Z',
+    ...commonArgs,
     isPrivate: true,
     numberOfFiles: 5,
+    imgSrc: '/coorong.png',
     isAdmin: true,
-    adminMenuItems: adminMenuItems,
-    handleClickFile: () => {
-      console.log('handle file click');
-    },
+    adminMenuItems,
+  },
+  argTypes: {
+    handleClickShare: { action: 'handle share click' },
+    handleClickFile: { action: 'handle file click' },
   },
   render: args => <PostCard {...args} signer={signer} />,
 };
+
+export const NoImage: Story = {
+  args: {
+    ...commonArgs,
+    isPrivate: true,
+    numberOfFiles: 5,
+    isAdmin: true,
+    adminMenuItems,
+  },
+  argTypes: {
+    handleClickShare: { action: 'handle share click' },
+    handleClickFile: { action: 'handle file click' },
+  },
+  render: args => <PostCard {...args} signer={signer} />,
+}
