+++ web-components/src/components/cards/card.stories.tsx
@@ -1,17 +1,12 @@
 import * as React from 'react';
 import { Box, Theme } from '@mui/material';
-import ReactHtmlParser from 'html-react-parser';
 import { makeStyles } from 'tss-react/mui';
 import { Track } from 'web-marketplace/src/lib/tracker/types';
-
-import { parseText } from 'src/utils/textParser';
 
 import { QuestionItem } from '../faq/Question';
 import CurrentCreditsIcon from '../icons/CurrentCreditsIcon';
 import FarmerIcon from '../icons/FarmerIcon';
 import TrustIcon from '../icons/TrustIcon';
-import { Body } from '../typography';
-import { User } from '../user/UserInfoCard';
 import Card from './Card';
 import GlanceCard from './GlanceCard';
 import GreenCard from './GreenCard';
@@ -21,7 +16,6 @@ import ImpactCard from './ImpactCard';
 import MapCard from './MapCard';
 import OnBoardingCard from './OnBoardingCard';
 import { OverviewCard } from './OverviewCard';
-import PostCard from './PostCard/PostCard';
 import ProjectCard from './ProjectCard';
 import ProjectImpactCard from './ProjectImpactCard/ProjectImpactCard';
 import PurchasedCreditsCard from './PurchasedCreditsCard';
