+++ web-components/src/components/icons/LockIcon.tsx
@@ -1,16 +1,27 @@
 import React from 'react';
+import SvgIcon, { SvgIconProps } from '@mui/material/SvgIcon';
 
-export const LockIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg width="24" height="24" fill="none" {...props}>
-    <path
-      d="M13 17.7324C13.5978 17.3866 14 16.7403 14 16C14 14.8954 13.1046 14 12 14C10.8954 14 10 14.8954 10 16C10 16.7403 10.4022 17.3866 11 17.7324V19C11 19.5523 11.4477 20 12 20C12.5523 20 13 19.5523 13 19V17.7324Z"
-      fill="currentColor"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M6 6V10H4C3.44772 10 3 10.4477 3 11V23C3 23.5523 3.44772 24 4 24H20C20.5523 24 21 23.5523 21 23V11C21 10.4477 20.5523 10 20 10H18V6C18 2.68629 15.3137 0 12 0C8.68629 0 6 2.68629 6 6ZM16 10V6C16 3.79086 14.2091 2 12 2C9.79086 2 8 3.79086 8 6V10H16ZM5 22V12H19V22H5Z"
-      fill="currentColor"
-    />
-  </svg>
+import { sxToArray } from '../../utils/mui/sxToArray';
+
+export const LockIcon = ({ sx, ...props }: SvgIconProps): JSX.Element => (
+  <SvgIcon
+    width="24"
+    height="24"
+    fill="none"
+    sx={[...sxToArray(sx)]}
+    {...props}
+  >
+    <svg>
+      <path
+        d="M13 17.7324C13.5978 17.3866 14 16.7403 14 16C14 14.8954 13.1046 14 12 14C10.8954 14 10 14.8954 10 16C10 16.7403 10.4022 17.3866 11 17.7324V19C11 19.5523 11.4477 20 12 20C12.5523 20 13 19.5523 13 19V17.7324Z"
+        fill="currentColor"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M6 6V10H4C3.44772 10 3 10.4477 3 11V23C3 23.5523 3.44772 24 4 24H20C20.5523 24 21 23.5523 21 23V11C21 10.4477 20.5523 10 20 10H18V6C18 2.68629 15.3137 0 12 0C8.68629 0 6 2.68629 6 6ZM16 10V6C16 3.79086 14.2091 2 12 2C9.79086 2 8 3.79086 8 6V10H16ZM5 22V12H19V22H5Z"
+        fill="currentColor"
+      />
+    </svg>
+  </SvgIcon>
 );
